requirements.txt
# ================================================================
# Google Calendar Automation - Dependencies
# ================================================================
# Python 3.8+ required
#
# Installation: pip install -r requirements.txt
# Update all: pip install --upgrade -r requirements.txt
# ================================================================

# Google API Libraries
# ================================================================
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Data Processing
# ================================================================
pandas==2.2.0
numpy==1.26.3

# Environment Management
# ================================================================
python-dotenv==1.0.1

# Additional Utilities (Optional but Recommended)
# ================================================================
# Uncomment if needed:
# openpyxl==3.1.2          # Excel file support
# requests==2.31.0         # HTTP requests
# pytz==2024.1             # Timezone handling
